+++ src/timer/dummy/SDL_systimer.c
@@ -65,7 +65,7 @@ SDL_GetPerformanceFrequency(void)
 }
 
 void
-SDL_Delay(Uint32 ms)
+SDL_Delay(double ms)
 {
     SDL_Unsupported();
 }
